src/backend/requirements/production.txt
fastapi
#sqlmodel
#alembic
uvicorn
python-dotenv
pydantic
pillow
pycairo
#psycopg
jinja2
